Inspector: attempt to detect and fix double formatting
Once https://github.com/canalplus/rx-player/pull/1469 is merged, we risk
to be in a situation where the `timestamp [namespace] log` format as
exported by the RxPaired's inspector (and can be imported by it) would
be unnecessarily repeating.

This is a proposal for fixing it by updating the inspector's code so we
can detect and just keep the first timestamp+namespace combo

To reduce the possibility of false positives and especially to avoid
loss of information, I check that the "namespace" are the same.

False positives are still a possibility though.