Fallback to HTTP POST when WebSocket aren't available
We encounter several devices on which WebSocket are not available / are
blocked.

This commit implements a fallback to HTTP POST-based log reporting when
the WebSocket connection establishement does not succeed.

The idea for now is that basically the client send logs to the RxPaired
server through HTTP POST every 2/3 seconds.

To divide messages sent in bulk between one another, I chose for now a
NULL byte (`\0`). This is kind of ugly but thankfully having a real NULL
byte in a sent log should be extremely rare.
I could also have JSONified everything and set logs inside an array.

This adds many complexities especially on the server, as now the
lifetime of a token might be unclear: is the device still using its
token or not? With a WebSocket we just checked the connection state (and
did keepalive ping-pong messages), when based on HTTP POST, we now rely
on a timeout without any message, for now set to 30 seconds.

For now it's not possible to use `sendInstruction` when a device relies
on HTTP POST. It could theoretically be done through several means
including Server-Sent-Events and long polling, but I did not bother for
now.