server: Fix log file being created each time an HTTP POST is received
We recently added the possibility to handle HTTP POST connections for
when WebSocket is not available.

However when doing that, we didn't think about the log files that may be
created by the server on each connection (the log filename including the
timestamp at the time of the connection), containing the logs sent by
the device.

It made sense with WebSocket to create one log file per connection, but
with HTTP POST, there might be a huge amount of HTTP POST performed by a
single device, leading to a huge amount of log files where we would
prefer a single one.

To fix this, I added the `firstConnectionTimestamp` metadata which is
the value of `Date.now()` on the first time an HTTP POST request has
been received from a device for the corresponding token.

Then, when deciding of the log filename, I look at if the device
connected to the corresponding token is connected through HTTP POST:
  - if it's the case, I rely on its `firstConnectionTimestamp`
  - if it's not the case (so probably WebSocket), I rely on `Date.now()`

To be sure the code is easier to maintain (to not mistakenly break it in
the future), I chose voluntarily to explicitly provide components of the
token's metadata that are required instead of the full `TokenMetadata`
object - which may at this point be still incomplete.