Do not rely on npmrc as a source for config anymore
npm now begins to stdout deprecation notices about relying on npmrc
files for configuration.

So I chose to rely on our own format of configuration file instead, (in
JSON).
I chose JSON due to the familiarity and simplicity of the format, but
one huge issue with JSON as a configuration format is the fact that it
does not support comments to document the various properties.

I tried to be smart with it with specific properties for comments, not
sure if it's readable that way though.